Added DNS lookup for Running VMs
Added a DNS lookup for Running VMs instead of relying on the Xymon hosts.cfg file to give me the name or on the VM name which isn't necessarily the DNS name. Xymon gets confused when the name in the hosts.cfg file doesn't match the DNS name of the server being monitored. Also cleaned up some spacing & indenting.